Move Search Form into separate component
