+++ apps/core/components/searchForm/PlacePicker/PlacePickerRenderer.js
