+++ apps/core/components/searchForm/PlacePicker/__tests__/helpers.test.js
